// src/types/openApi.ts
export interface IDiscriminator {
  propertyName: string;
  mapping?: {
    [k: string]: string;
  };
}

export interface IEncoding {
  contentType?: string;
  headers?: {
    [k: string]: THeader;
  };
  style?: 'form' | 'spaceDelimited' | 'pipeDelimited' | 'deepObject';
  explode?: boolean;
  allowReserved?: boolean;
}

export interface IExample {
  summary?: string;
  description?: string;
  value?: any;
  externalValue?: string;

  [k: string]: any;
}

export interface IExternalDocumentation {
  description?: string;
  url: string;

  [k: string]: any;
}
// ...
export interface IReference {
  $ref: string;
}
// ...
export interface ISchema {
  // The following properties are taken directly from the JSON Schema definition and follow the same specifications
  title?: string;
  multipleOf?: number;
  maximum?: number;
  exclusiveMaximum?: boolean;
  minimum?: number;
  exclusiveMinimum?: boolean;
  maxLength?: number;
  minLength?: number;
  pattern?: string;
  maxItems?: number;
  minItems?: number;
  uniqueItems?: boolean;
  maxProperties?: number;
  minProperties?: number;
  required?: string[];
  enum?: any[];
  // The following properties are taken from the JSON Schema definition but their definitions were adjusted to the OpenAPI Specification
  type?: 'array' | 'boolean' | 'integer' | 'number' | 'object' | 'string';
  not?: ISchema | IReference;
  allOf?: Array<ISchema | IReference>;
  oneOf?: Array<ISchema | IReference>;
  anyOf?: Array<ISchema | IReference>;
  items?: ISchema | IReference;
  properties?: {
    [k: string]: ISchema | IReference;
  };
  propertyNames?: ISchema | IReference;
  additionalProperties?: ISchema | IReference | boolean;
  description?: string;
  format?: string;
  default?: any;
  nullable?: boolean;
  discriminator?: IDiscriminator;
  readOnly?: boolean;
  writeOnly?: boolean;
  example?: any;
  externalDocs?: IExternalDocumentation;
  deprecated?: boolean;
  xml?: IXML;

  [k: string]: any;
}
// ...
export interface IXML {
  name?: string;
  namespace?: string;
  prefix?: string;
  attribute?: boolean;
  wrapped?: boolean;

  [k: string]: any;
}